debundle/lowering: extract runtime-source URL rewriter
BAZEL_TEST_INVOCATIONS=none: file move, no semantic change.
